Fix: 댓글 한번에 달리는 오류 해결
