+++ src/pages/seongeunPark/Main/Main.scss
@@ -332,8 +332,6 @@
 .input-comments {
   width: 80%;
   background-color: transparent;
-  color: E2E2E2;
-  opacity: 0.6;
 }
 
 .comments-btn {
